Use one port per container
Simplify the structure of the port by mapping 1 port to 1 container.
In the old model, several containers could be run from one port.

The container structure is now allocated when the port starts. The
liblxc API for freeing the container structure is not public so the
container structure must exist for the life of the port. Remove
the new/2 message type.

Clarify the generated erlang code by using Container as the variable
holding the pid. However, doesn't make sense in some places still, i.e.,
for operations concerning all containers like list/3.