Normalize the return values
Follow the liblxc API and return a binary when a function returns char *
and a bool when the function returns bool.

For get functions, on error, either an empty binary is returned or the
atom 'none'. This should be normalized in the future too.

The liblxc API has provisions for returning an error string/number but
nothing seems to set it. So avoid guessing the error values.