Disable stdout before running any lxc function
The create function templates may write to stdout, interfering with the
port communication. As a precaution, redirect any writes to standard
output to null before running any liblxc function.

This causes other problems: a simple container can be run without a tty,
using stdin for input and stdout for output. This is useful when
building a pipeline of sandboxed commands or running a sandboxed service
and having all the output handled by the unsandboxed erlang side.