Add spawn_link/0,1,2
Add a spawn_link/0,1,2 corresponding to spawn/0,1,2. The semantics are
not correct yet. The idea is that if a container crashes, the owner will
not crash. However, the owner will probably crash anyways when it
queries the container state or tries to send it a message on the
console.

spawn_link/0,1,2 will also crash when the port exits. The container will
keep on running when beam exits though.