Remove daemonize/2, always use daemonize
Reduce the amount of code by removing support for forking the on start.
Originally the plan was:

* have the port multiplex many containers

The port was simplified to be one port per container.

* support containerized processes that communicated via stdin/stdout

This could be supported by switching to 'stream' mode and adding some
framing. We could use inband signalling to switch from "command" mode to
"input" mode. But it would be simpler to write a new stream based port.
Something like:

* check if tty's are defined for the container

* if not (or if specified through an option), have erlxc_console spawn
  an erlxc_console process

    erlxc_console -n <name> -p <path>

* the port calls lxc_start