Notifications for state changes
liblxc uses an abstract socket to poll for state changes. Add a
simple mechanism for notifying the erlang process of state changes.
Doesn't appear to be a way of subscribing to events, e.g., selecting on
the command socket for the container.

For now, a Unix process polls the sockets at an interval specified by
the user (in seconds). The poll process will also check every time a
message is sent to the port.