liblxc: crash on bad argument
Move error from erxcl_drv to liblxc so the appropriate arguments are
returned to the user. As a consequence, erlxc_drv will no longer crash
when a bad argument is passed but will pass through the atom 'badarg'
This is the appropriate behaviour since the port is not in an
inconsistent state after a badarg.