imc: a read-only container
Build a chroot and run an unprivileged command inside the container.
Read-only directories are bind mounted from the host. Writable directories
are mounted as tmpfs.

The command is run under a random UID/GID. stdout/stderr is not reported
back to erlang, so the only way to communicate with the container is
via the network.

By default, a distributed erlang node is run. To start the node, compile
the example:

    make eg

Then run a node:

    ./start.sh
    % node name will be: imc@192.168.213.45
    1> Container = imc:spawn("192.168.123.45").

Start up another distributed erlang node and connect:

    % Assuming 192.168.213.10 is the IP address of your client
    erl -name c@192.168.123.10 -setcookie COOKIE
    ^G
    (c@192.168.123.10)1>
    User switch command
     --> r 'imc@192.168.123.45'
     --> c
    Eshell V5.8.5  (abort with ^G)
    (imc@192.168.123.45)1>