erlxc: generate the version from the app file
Use an escript to get the version number from the app file for use in
the erlxc executable. Use file:consult/1 instead of application/load/1
because the final app file is generated by rebar and so won't exist
before the compile phase is run.

All these relative paths are sketchy and may break including erlxc as a
dependency (needs to be tested).