erlxc: generate the container name from a template
Allow more flexible naming of the container by adding supports for
name templates.

Templates must begin with "@". Any "#"'s in the name are replaced with
a random alphanumeric character. The "@" is not included in the
container name:

    erlxc:spawn("@foo######")

Unlike named containers, containers that use a template are marked as
being temporary (destroyed when the erlxc process exits).

    erlxc:spawn("foo123")  % transient
    erlxc:spawn("@foo123") % temporary

Containers work as follows:

    erlxc:spawn()
    erlxc:spawn(<<>>)
    erlxc:spawn(<<"@">>)

        Create a temporary container named using the default template
        ("erlxc########")

    erlxc:spawn("@abc###")

        Create a temporary container using the specified template.

    erlxc:spawn("abc###")

        Create a transient container literally named "abc###".