erlxc:call/3: return badarg on failure
erlxc:call/3 is a wrapper around liblxc:call/3 for boolean functions
that crashes on error (false values). In the case of an error, likely a
bad argument has been passed in, e.g., an invalid bridge interface.

Note badarg is overloaded here: the port will also return badarg if an
argument of an inappropriate type is used.